+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Next
 * Improve the spinner while waiting for ``minidcos`` commands.
 * Add ``send-file`` commands to ``minidcos`` subcommands.
 * Remove ``--env`` option on ``minidcos docker inspect``.
+* Custom backends must now specify the base config in the ``ClusterBackend`` rather than the ``ClusterManager``.
 
 2019.04.08.1
 ------------